cleanup: redundant type casts

+++ src/algvis/ds/priorityqueues/lazybinomialheap/LazyBinomialHeap.java
@@ -57,8 +57,8 @@ public class LazyBinomialHeap extends BinomialHeap {
     @Override
     public void meld(int i, int j) {
         final Pair<Integer, Integer> p = chooseHeaps(i, j);
-        i = (int)p.first;
-        j = (int)p.second;
+        i = p.first;
+        j = p.second;
         ((MeldablePQButtons) panel.buttons).activeHeap.setValue(i);
         start(new LazyBinHeapMeld(this, i, j));
     }
