+++ src/algvis/ds/priorityqueues/leftistheap/LeftHeap.java
@@ -78,8 +78,8 @@ public class LeftHeap extends MeldablePQ implements ClickListener {
     @Override
     public void meld(int i, int j) {
         final Pair<Integer, Integer> p = chooseHeaps(i, j);
-        i = (int)p.first;
-        j = (int)p.second;
+        i = p.first;
+        j = p.second;
         ((MeldablePQButtonsNoDecr) panel.buttons).activeHeap.setValue(i);
         start(new LeftHeapMeld(this, i, j));
     }
