+++ src/algvis/ds/priorityqueues/pairingheap/PairingHeap.java
@@ -92,8 +92,8 @@ public class PairingHeap extends MeldablePQ implements ClickListener {
     @Override
     public void meld(int i, int j) {
         final Pair<Integer, Integer> p = chooseHeaps(i, j);
-        i = (int)p.first;
-        j = (int)p.second;
+        i = p.first;
+        j = p.second;
         ((MeldablePQButtons) panel.buttons).activeHeap.setValue(i);
         start(new PairHeapMeld(this, i, j));
     }
